requirements.txt
openai>=1.0.0
fastapi>=0.115.0
uvicorn>=0.24.0
pydantic>=2.0.0
websockets>=12.0
gymnasium>=0.29.0
openenv-core>=0.1.0
